Refactor CSV export to use a reusable download function and update BOM option
